containers/geoparse/requirements.txt
-e git://github.com/guma44/GEOparse.git#egg=GEOparse
biopython
pyfunctional
click